+++ docs/progress/dpi.md
@@ -35,25 +35,25 @@ Two frontiers are open. On the C++ backend the function import surface (D1-D3) i
 marshaling surface (D4, D4b) runs -- an export executes end to end through the import -> export
 chain across every argument direction and both value families (by-value scalars, by-pointer scalar
 `output` / `inout`, and canonical 2-state / 4-state packed vectors), its wrapper recovering the
-exported subroutine's single top-level instance from the running design -- and the DPI task surface
-runs in both directions: the import task (D5) has SV call a foreign C task, and the export task (D6)
-has foreign C call back an SV task, both under the uniform task protocol. A foreign task also
-consumes simulation time (D6b): its call runs on a fiber, so an exported task it reaches suspends on
-a delay across the boundary and resumes when time advances. The `svdpi` context surface (D7) runs: a
+exported subroutine's receiver from the running design -- and the DPI task surface runs in both
+directions: the import task (D5) has SV call a foreign C task, and the export task (D6) has foreign
+C call back an SV task, both under the uniform task protocol. A foreign task also consumes
+simulation time (D6b): its call runs on a fiber, so an exported task it reaches suspends on a delay
+across the boundary and resumes when time advances. The `svdpi` context surface (D7) runs: a
 `context` import observes the fully qualified scope of its declaration, foreign code sets and gets
 the current scope and resolves a scope to and from its name, stores per-scope user data, and reads
 the scope's effective time unit and precision and the current time scaled to it -- the current scope
 riding the calling process so concurrent time-consuming imports never cross. Instance-bound export
 dispatch (D4a) now runs too: a module-scoped export dispatches to the instance the foreign call
 targets -- the current DPI scope the caller established, or one `svSetScope` redirected to -- rather
-than a single hardcoded instance, and a receiver-less package-scoped export calls its package
-function directly with no instance; reaching any export now requires a valid scope context, so its
-caller must be a context import or set the scope with `svSetScope`. What remains is `$unit`-scoped
-export (D4c, blocked on the `$unit` scope itself), the generated ABI header and export linkage (D9),
-and the disable protocol across the boundary (D6c). On the execution backend scalar import (D10) is
-in: a foreign call lowers to an external-linkage symbol and the by-value carriers marshal. The rest
-of the import surface (D11) is blocked, and not by anything DPI owns: by-pointer marshaling is
-expressed as a closure, which that backend does not yet lower at all (`architecture-reset.md`).
+than a single hardcoded instance, and a receiver-less package- or `$unit`-scoped export calls its
+free function directly with no instance (D4a, D4c); reaching any export now requires a valid scope
+context, so its caller must be a context import or set the scope with `svSetScope`. What remains is
+the generated ABI header and export linkage (D9), and the disable protocol across the boundary
+(D6c). On the execution backend scalar import (D10) is in: a foreign call lowers to an
+external-linkage symbol and the by-value carriers marshal. The rest of the import surface (D11) is
+blocked, and not by anything DPI owns: by-pointer marshaling is expressed as a closure, which that
+backend does not yet lower at all (`architecture-reset.md`).
 
 ## Sub-Steps
 
@@ -111,10 +111,10 @@ from an external main.
       scope with `svSetScope` (LRM 35.5.3). This is the `mhpmcounter_num` / `mhpmcounter_get` shape
       in the Ibex bring-up; the pure-SV Ibex run never calls them, so this closes the construct, not
       the Ibex external-driver usage.
-- [ ] D4c -- `$unit`-scoped receiver-less export (LRM 35.7): the same package-export wrapper shape,
-      but the exported subroutine lives in the anonymous compilation-unit scope. Blocked on the
-      `$unit` scope itself being modeled as a unit (`packages.md` PK5); once it is, its export hangs
-      off it exactly as D4a's package export does.
+- [x] D4c -- `$unit`-scoped receiver-less export (LRM 35.7): a subroutine at compilation-unit scope
+      (LRM 3.12.1) exports exactly as a package function does. The `$unit` scope is an anonymous
+      namespace unit, so its export rides the same receiver-less wrapper as D4a's package export
+      with no added machinery.
 - [x] D4b -- 4-state and by-pointer packed export marshaling. An `output` / `inout` scalar crosses
       by pointer to its by-value carrier; a packed vector of any direction crosses by pointer as its
       canonical `svBitVecVal*` / `svLogicVecVal*` buffer, its planes reshaped without a per-bit
