+++ src/lyra/backend/cpp/render_call.cpp
@@ -296,8 +296,6 @@ auto BuiltinFnCppName(support::BuiltinFn id) -> std::string_view {
       return "RunExportedTaskToCompletion";
     case support::BuiltinFn::kCurrentExportScope:
       return "CurrentExportScope";
-    case support::BuiltinFn::kCurrentExportEffects:
-      return "CurrentExportEffects";
     case support::BuiltinFn::kFromSvLogic:
       return "FromSvLogic";
     case support::BuiltinFn::kFromInt:
@@ -419,7 +417,6 @@ auto BuiltinFnCppNamespace(support::BuiltinFn id) -> std::string_view {
     case support::BuiltinFn::kValuePlusargs:
     case support::BuiltinFn::kRunExportedTaskToCompletion:
     case support::BuiltinFn::kCurrentExportScope:
-    case support::BuiltinFn::kCurrentExportEffects:
       return "lyra::runtime";
     default:
       return "";
